+++ src/proxy_protect/components/webserver.clj
@@ -1,8 +1,7 @@
 (ns proxy-protect.components.webserver
   (:require [clojure.tools.logging :refer [info error]]
             [com.stuartsierra.component :as component]
-            [ring.adapter.jetty :refer [run-jetty]])
-  (:import org.eclipse.jetty.server.Server))
+            [org.httpkit.server :as server]))
 
 
 (defrecord WebServer [handler
@@ -12,21 +11,16 @@
   (start [component]
     (info "Starting Webserver Component...")
     (let [rules (:rules rules)]
-      (assoc component :webserver
-                       (run-jetty (handler rules)
-                                  {:host "0.0.0.0"
-                                   :port 3000
-                                   :join? false}))))
+      (assoc component :webserver (server/run-server (handler rules)
+                                                     {:host "0.0.0.0"
+                                                      :port 3000}))))
   (stop [component]
     (info "Stopping Webserver Component...")
     (try
-      (if-let [webserver (:webserver component)]
-        (.stop ^Server webserver)
-        (.destroy ^Server webserver))
+      (when webserver
+        (webserver))
       (catch Exception e
-        (error e (str "Error stopping Webserver"))
-        (assoc component :handler nil
-                         :webserver nil)))
+        (error e (str "Error stopping Webserver"))))
     (info "Stopped Webserver Component.")
     (assoc component :handler nil
                      :rules nil